perf(admin-analytics): cache snapshot via daily scheduler; drop unbounded reads from hot path
`adminAnalytics` was the single largest line item on the Cloud Functions
invoice (~$5/10 days) once the user base passed a handful of teachers.
Every admin page load triggered two unbounded Firestore reads:
`collectionGroup('dashboards').stream()` and
`collection('ai_usage').stream()`, on a 4 GiB / 540 s instance.

The compute is now a scheduled job — `recomputeAdminAnalytics` runs at
5 AM America/Chicago daily, iterates non-archived orgs, and writes
the payload to `/organizations/{orgId}/analytics/snapshot` with
`computedAt` and `nextRecomputeAt` timestamps. The HTTP `adminAnalytics`
handler now reads that snapshot doc and returns it verbatim. Hot-path
memory drops from 4 GiB to 1 GiB, timeout from 540 s to 30 s, and the
two streaming reads are eliminated from the per-call path entirely.

Cold-start contract: a brand-new org with no snapshot yet returns 503
with `error: 'not-yet-computed'` and a user-facing "ready by 5:00 AM
Central" message. The UI's Refresh button is gone — analytics are
explicitly a luxury daily read, and a manual recompute button would
defeat the cache. Replaced with a "Updated 4h ago · Next update at
5:00 AM" freshness badge that surfaces the cached nature of the data
so admins understand they're not looking at a live aggregate.

Layout:
- `functions/src/adminAnalyticsCompute.ts` — pure compute helper, owns
  the two unbounded reads.
- `functions/src/adminAnalyticsSnapshot.ts` — scheduled `onSchedule`
  trigger + `readAnalyticsSnapshot()` helper used by the hot path.
  Per-org compute failures are caught and logged; one bad org doesn't
  starve the rest of the batch.
- `functions/src/index.ts` — adminAnalytics handler now ~50 LOC
  (auth/authz + snapshot read), down from ~580 LOC of inline compute.

Firestore rules: added `match /organizations/{orgId}/analytics/{document}`
gated to org admins (super/domain/building). Writes blocked — the
Admin SDK bypass on the scheduler is the sole writer.

Tests: existing compute tests now call `computeAnalyticsForOrg`
directly instead of going through the HTTP handler boilerplate; added
new coverage for the snapshot-read hot path (cold-start 503, cached
200 with meta) and the scheduled recompute (status filter, per-org
failure resilience). All 208 functions tests pass; 2301 root tests
unchanged.

Deploy note: the first scheduler run is at 5 AM Central on the day
after deploy. Until then, admins will see the cold-start "not yet
computed" message. To populate snapshots immediately, manually trigger
`recomputeAdminAnalytics` from the GCP console after deploy.

Closes the #1 action item from the cost audit at
`docs/CLOUD_FUNCTIONS_COST_OPTIMIZATION.md`.
