+++ functions/src/index.test.ts
@@ -32,6 +32,17 @@ const mockFirestoreState = {
   users: [] as MockDocInput[],
   dashboards: [] as MockDocInput[],
   aiUsage: [] as MockDocInput[],
+  // Map of path → doc data for `db.doc(path).get()` mocks. Used by the
+  // adminAnalytics snapshot read tests (path:
+  // `organizations/{orgId}/analytics/snapshot`) and any future test that
+  // needs a deterministic doc fetch. Unset paths return `{ exists: false }`
+  // via the default `mockFirestore.doc` fallback.
+  docs: new Map<string, Record<string, unknown>>(),
+  // Org records for the scheduled `recomputeAdminAnalytics` job. Each entry
+  // is treated as a doc under `/organizations/{id}` with the given `status`.
+  // The scheduler iterates this collection and only recomputes for orgs
+  // with status in {active, trial}.
+  organizations: [] as { id: string; status: string }[],
 };
 
 const toDocSnapshot = (doc: MockDocInput) => ({
@@ -61,7 +72,17 @@ interface MockDocRef {
 const mockFirestore = {
   doc: vi.fn((path: string) => ({
     path,
-    get: vi.fn(() => Promise.resolve({ exists: false })),
+    get: vi.fn(() => {
+      const stashed = mockFirestoreState.docs.get(path);
+      if (stashed) {
+        return Promise.resolve({ exists: true, data: () => stashed });
+      }
+      return Promise.resolve({ exists: false });
+    }),
+    set: vi.fn((data: Record<string, unknown>) => {
+      mockFirestoreState.docs.set(path, data);
+      return Promise.resolve();
+    }),
   })),
   getAll: vi.fn((...refs: MockDocRef[]) => {
     return Promise.resolve(
@@ -134,6 +155,21 @@ const mockFirestore = {
         select: vi.fn(() => ({
           stream: vi.fn(() => toAsyncStream(mockFirestoreState.aiUsage)),
         })),
+      };
+    }
+
+    // Top-level `organizations` collection — used by the scheduled
+    // `recomputeAdminAnalytics` job to iterate every non-archived org.
+    if (name === 'organizations') {
+      return {
+        get: vi.fn(() =>
+          Promise.resolve({
+            docs: mockFirestoreState.organizations.map((o) => ({
+              id: o.id,
+              data: () => ({ status: o.status }),
+            })),
+          })
+        ),
       };
     }
 
@@ -245,6 +281,13 @@ vi.mock('firebase-functions/v2', () => ({
   setGlobalOptions: vi.fn(),
 }));
 
+// Mock firebase-functions/v2/scheduler — `onSchedule` returns the handler
+// directly so tests can invoke the recompute job by calling
+// `recomputeAdminAnalytics()` like a plain async function.
+vi.mock('firebase-functions/v2/scheduler', () => ({
+  onSchedule: (_options: unknown, handler: () => Promise<void>) => handler,
+}));
+
 // Mock firebase-functions/params (defineSecret)
 vi.mock('firebase-functions/params', () => ({
   defineSecret: (name: string) => ({
@@ -263,6 +306,12 @@ import {
   adminAnalytics,
   getPseudonymsForAssignmentV1,
 } from './index';
+import { computeAnalyticsForOrg } from './adminAnalyticsCompute';
+import {
+  SNAPSHOT_SCHEMA_VERSION,
+  recomputeAdminAnalytics,
+  type AnalyticsSnapshotDoc,
+} from './adminAnalyticsSnapshot';
 
 describe('fetchExternalProxy', () => {
   beforeEach(() => {
@@ -497,6 +546,7 @@ describe('adminAnalytics', () => {
     mockFirestoreState.users = [];
     mockFirestoreState.dashboards = [];
     mockFirestoreState.aiUsage = [];
+    mockFirestoreState.docs = new Map();
   });
 
   it('aggregates users by domain and building, including no-building users', async () => {
@@ -546,32 +596,7 @@ describe('adminAnalytics', () => {
       { id: 'uid2_smart-poll_2026-03-30', data: { count: 99 } },
     ];
 
-    const handler = adminAnalytics;
-
-    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
-    const resPromise = new Promise<any>((resolve) => {
-      const mockRes = {
-        status: vi.fn().mockReturnThis(),
-        json: vi.fn().mockImplementation((data) => {
-          resolve(data);
-          return data;
-        }),
-        setHeader: vi.fn().mockReturnThis(),
-        getHeader: vi.fn().mockReturnValue(''),
-      };
-
-      const mockReq = {
-        headers: {
-          origin: 'http://localhost',
-          authorization: 'Bearer mock-token',
-        },
-        body: { orgId: 'orono' },
-      };
-
-      (handler as any)(mockReq, mockRes);
-    });
-
-    const capturedData = await resPromise;
+    const capturedData = await computeAnalyticsForOrg('orono');
 
     expect(capturedData.users.total).toBe(3);
     expect(capturedData.users.monthly).toBe(2);
@@ -597,7 +622,6 @@ describe('adminAnalytics', () => {
       daily: 0,
     });
     expect(capturedData.api.totalCalls).toBe(10);
-    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return */
   });
 
   it('returns topUsers with resolved email and unknown fallback', async () => {
@@ -624,32 +648,7 @@ describe('adminAnalytics', () => {
       { id: 'uid_b_2026-03-30', data: { count: 3 } },
     ];
 
-    const handler = adminAnalytics;
-
-    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
-    const resPromise = new Promise<any>((resolve) => {
-      const mockRes = {
-        status: vi.fn().mockReturnThis(),
-        json: vi.fn().mockImplementation((data) => {
-          resolve(data);
-          return data;
-        }),
-        setHeader: vi.fn().mockReturnThis(),
-        getHeader: vi.fn().mockReturnValue(''),
-      };
-
-      const mockReq = {
-        headers: {
-          origin: 'http://localhost',
-          authorization: 'Bearer mock-token',
-        },
-        body: { orgId: 'orono' },
-      };
-
-      (handler as any)(mockReq, mockRes);
-    });
-
-    const capturedData = await resPromise;
+    const capturedData = await computeAnalyticsForOrg('orono');
 
     expect(capturedData.api.topUsers[0]).toEqual({
       uid: 'uid_a',
@@ -661,7 +660,6 @@ describe('adminAnalytics', () => {
       count: 3,
       email: 'Unknown (uid_b)',
     });
-    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return */
   });
 
   it('returns usersByType with correct count and resolved emails', async () => {
@@ -712,28 +710,7 @@ describe('adminAnalytics', () => {
       },
     ];
 
-    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
-    const resPromise = new Promise<any>((resolve) => {
-      const mockRes = {
-        status: vi.fn().mockReturnThis(),
-        json: vi.fn().mockImplementation((data) => {
-          resolve(data);
-          return data;
-        }),
-        setHeader: vi.fn().mockReturnThis(),
-        getHeader: vi.fn().mockReturnValue(''),
-      };
-      const mockReq = {
-        headers: {
-          origin: 'http://localhost',
-          authorization: 'Bearer mock-token',
-        },
-        body: { orgId: 'orono' },
-      };
-      (adminAnalytics as any)(mockReq, mockRes);
-    });
-
-    const capturedData = await resPromise;
+    const capturedData = await computeAnalyticsForOrg('orono');
 
     // clock: 2 distinct users (alice counted once despite 2 dashboards)
     expect(capturedData.widgets.usersByType.clock.count).toBe(2);
@@ -750,7 +727,6 @@ describe('adminAnalytics', () => {
     expect(capturedData.widgets.usersByType.timer.emails).toEqual([
       'alice@district.org',
     ]);
-    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return */
   });
 
   it('caps usersByType emails at 20 and reports accurate count up to 100', async () => {
@@ -770,34 +746,12 @@ describe('adminAnalytics', () => {
       data: { updatedAt: Date.now(), widgets: [{ type: 'clock' }] },
     }));
 
-    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
-    const resPromise = new Promise<any>((resolve) => {
-      const mockRes = {
-        status: vi.fn().mockReturnThis(),
-        json: vi.fn().mockImplementation((data) => {
-          resolve(data);
-          return data;
-        }),
-        setHeader: vi.fn().mockReturnThis(),
-        getHeader: vi.fn().mockReturnValue(''),
-      };
-      const mockReq = {
-        headers: {
-          origin: 'http://localhost',
-          authorization: 'Bearer mock-token',
-        },
-        body: { orgId: 'orono' },
-      };
-      (adminAnalytics as any)(mockReq, mockRes);
-    });
-
-    const capturedData = await resPromise;
+    const capturedData = await computeAnalyticsForOrg('orono');
 
     // All 25 distinct users tracked (well within the 100-cap)
     expect(capturedData.widgets.usersByType.clock.count).toBe(25);
     // Emails preview capped at 20
     expect(capturedData.widgets.usersByType.clock.emails).toHaveLength(20);
-    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return */
   });
 
   it('excludes anonymous auth users from all analytics metrics', async () => {
@@ -853,28 +807,7 @@ describe('adminAnalytics', () => {
       { id: 'uid_anon1_2026-03-30', data: { count: 10 } },
     ];
 
-    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
-    const resPromise = new Promise<any>((resolve) => {
-      const mockRes = {
-        status: vi.fn().mockReturnThis(),
-        json: vi.fn().mockImplementation((data) => {
-          resolve(data);
-          return data;
-        }),
-        setHeader: vi.fn().mockReturnThis(),
-        getHeader: vi.fn().mockReturnValue(''),
-      };
-      const mockReq = {
-        headers: {
-          origin: 'http://localhost',
-          authorization: 'Bearer mock-token',
-        },
-        body: { orgId: 'orono' },
-      };
-      (adminAnalytics as any)(mockReq, mockRes);
-    });
-
-    const capturedData = await resPromise;
+    const capturedData = await computeAnalyticsForOrg('orono');
 
     // Only the teacher should be counted — anonymous users excluded
     expect(capturedData.users.total).toBe(1);
@@ -896,7 +829,6 @@ describe('adminAnalytics', () => {
 
     // Only teacher's AI usage counted
     expect(capturedData.api.totalCalls).toBe(5);
-    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
   });
 
   it('excludes signed-in auth users who are not members of the requested org', async () => {
@@ -977,28 +909,7 @@ describe('adminAnalytics', () => {
       { id: 'uid_uninvited_same_domain_2026-03-30', data: { count: 42 } },
     ];
 
-    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
-    const resPromise = new Promise<any>((resolve) => {
-      const mockRes = {
-        status: vi.fn().mockReturnThis(),
-        json: vi.fn().mockImplementation((data) => {
-          resolve(data);
-          return data;
-        }),
-        setHeader: vi.fn().mockReturnThis(),
-        getHeader: vi.fn().mockReturnValue(''),
-      };
-      const mockReq = {
-        headers: {
-          origin: 'http://localhost',
-          authorization: 'Bearer mock-token',
-        },
-        body: { orgId: 'orono' },
-      };
-      (adminAnalytics as any)(mockReq, mockRes);
-    });
-
-    const capturedData = await resPromise;
+    const capturedData = await computeAnalyticsForOrg('orono');
 
     // Only the invited member counts toward totals.
     expect(capturedData.users.total).toBe(1);
@@ -1034,7 +945,6 @@ describe('adminAnalytics', () => {
       (u: { email: string }) => u.email === 'uninvited@school.org'
     );
     expect(uninvitedRow).toBeUndefined();
-    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
   });
 
   it('counts invited-but-never-signed-in members toward totals with zero engagement', async () => {
@@ -1072,28 +982,7 @@ describe('adminAnalytics', () => {
       },
     ];
 
-    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
-    const resPromise = new Promise<any>((resolve) => {
-      const mockRes = {
-        status: vi.fn().mockReturnThis(),
-        json: vi.fn().mockImplementation((data) => {
-          resolve(data);
-          return data;
-        }),
-        setHeader: vi.fn().mockReturnThis(),
-        getHeader: vi.fn().mockReturnValue(''),
-      };
-      const mockReq = {
-        headers: {
-          origin: 'http://localhost',
-          authorization: 'Bearer mock-token',
-        },
-        body: { orgId: 'orono' },
-      };
-      (adminAnalytics as any)(mockReq, mockRes);
-    });
-
-    const capturedData = await resPromise;
+    const capturedData = await computeAnalyticsForOrg('orono');
 
     // Totals include the invited member.
     expect(capturedData.users.total).toBe(2);
@@ -1117,9 +1006,10 @@ describe('adminAnalytics', () => {
 
     // Per-user drilldown: the invited member is present with zero engagement.
     const invitedRow = capturedData.users.userList.find(
-      (u: { email: string }) => u.email === 'invited@district.org'
+      (u) => u.email === 'invited@district.org'
     );
     expect(invitedRow).toBeDefined();
+    if (!invitedRow) throw new Error('unreachable');
     expect(invitedRow.lastSignInMs).toBe(0);
     expect(invitedRow.lastEditMs).toBe(0);
     expect(invitedRow.hasDashboard).toBe(false);
@@ -1129,7 +1019,6 @@ describe('adminAnalytics', () => {
     // Only the active member owns a dashboard.
     expect(capturedData.users.withDashboards).toBe(1);
     expect(capturedData.dashboards.total).toBe(1);
-    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
   });
 
   it('returns 400 when orgId is missing from the request body', async () => {
@@ -1336,6 +1225,274 @@ describe('adminAnalytics', () => {
       collectionSpy.mockRestore();
     }
     /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment */
+  });
+
+  it('returns 503 when no analytics snapshot exists yet for the org', async () => {
+    // Cold-start contract: a brand-new org has no entry at
+    // `organizations/{orgId}/analytics/snapshot` until the scheduled
+    // `recomputeAdminAnalytics` job has run at least once. The HTTP handler
+    // must return a deterministic 503 with `error: 'not-yet-computed'` so
+    // the UI can show "Analytics ready after the next scheduled refresh"
+    // rather than spinning forever or surfacing a generic error.
+    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-call */
+    const statusSpy = vi.fn().mockReturnThis();
+    const jsonSpy = vi.fn().mockReturnThis();
+    const setHeaderSpy = vi.fn().mockReturnThis();
+    const mockRes = {
+      status: statusSpy,
+      json: jsonSpy,
+      setHeader: setHeaderSpy,
+      getHeader: vi.fn().mockReturnValue(''),
+    };
+
+    const mockReq = {
+      headers: {
+        origin: 'http://localhost',
+        authorization: 'Bearer mock-token',
+      },
+      body: { orgId: 'orono' },
+    };
+
+    await (adminAnalytics as any)(mockReq, mockRes);
+
+    expect(statusSpy).toHaveBeenCalledWith(503);
+    const jsonCall = jsonSpy.mock.calls[0]?.[0] as
+      | { error?: string; message?: string }
+      | undefined;
+    expect(jsonCall?.error).toBe('not-yet-computed');
+    expect(jsonCall?.message ?? '').toMatch(/scheduled refresh/i);
+
+    // Crucially: the snapshot-only hot path must not have hit either of the
+    // unbounded streaming reads that the old inline-compute implementation
+    // did. Verifies the cost-saving invariant — if a regression reintroduces
+    // the streams, this assertion catches it.
+    expect(mockFirestore.collectionGroup).not.toHaveBeenCalled();
+    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-call */
+  });
+
+  it('returns the cached snapshot payload + meta when one exists', async () => {
+    // Seed a snapshot doc at the canonical path. The hot path should read
+    // this verbatim and round-trip the payload to the caller alongside the
+    // freshness metadata the UI badge consumes.
+    const computedAt = Date.now() - 60 * 60 * 1000; // 1 hour ago
+    const nextRecomputeAt = computedAt + 24 * 60 * 60 * 1000;
+    const seedSnapshot: AnalyticsSnapshotDoc = {
+      schemaVersion: SNAPSHOT_SCHEMA_VERSION,
+      computedAt,
+      nextRecomputeAt,
+      computeDurationMs: 42_000,
+      payload: {
+        users: {
+          total: 7,
+          registered: 7,
+          registeredIsFallback: false,
+          monthly: 5,
+          daily: 2,
+          withDashboards: 6,
+          domains: { 'district.org': { total: 7, monthly: 5, daily: 2 } },
+          buildings: { north: { total: 7, monthly: 5, daily: 2 } },
+          domainBuilding: {},
+          userList: [],
+        },
+        widgets: { totalInstances: {}, activeInstances: {}, usersByType: {} },
+        dashboards: { total: 6, avgWidgetsPerDashboard: 3.5 },
+        api: {
+          totalCalls: 123,
+          activeUsers: 4,
+          topUsers: [],
+          avgDailyCalls: 41,
+          avgDailyCallsPerUser: 10.25,
+          byFeature: {},
+        },
+      },
+    };
+    mockFirestoreState.docs.set(
+      'organizations/orono/analytics/snapshot',
+      seedSnapshot as unknown as Record<string, unknown>
+    );
+
+    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-call */
+    let captured: unknown;
+    const statusSpy = vi.fn().mockReturnThis();
+    const jsonSpy = vi.fn().mockImplementation((data: unknown) => {
+      captured = data;
+      return data;
+    });
+    const mockRes = {
+      status: statusSpy,
+      json: jsonSpy,
+      setHeader: vi.fn().mockReturnThis(),
+      getHeader: vi.fn().mockReturnValue(''),
+    };
+
+    const mockReq = {
+      headers: {
+        origin: 'http://localhost',
+        authorization: 'Bearer mock-token',
+      },
+      body: { orgId: 'orono' },
+    };
+
+    await (adminAnalytics as any)(mockReq, mockRes);
+    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-call */
+
+    // 200 (no explicit status call — Express defaults to 200 when only
+    // `res.json` is invoked).
+    expect(statusSpy).not.toHaveBeenCalledWith(503);
+    expect(statusSpy).not.toHaveBeenCalledWith(500);
+
+    const body = captured as {
+      users?: { total?: number };
+      dashboards?: { total?: number };
+      meta?: { computedAt?: number; nextRecomputeAt?: number };
+    };
+    expect(body.users?.total).toBe(7);
+    expect(body.dashboards?.total).toBe(6);
+    expect(body.meta?.computedAt).toBe(computedAt);
+    expect(body.meta?.nextRecomputeAt).toBe(nextRecomputeAt);
+
+    // No streaming reads on the cached path.
+    expect(mockFirestore.collectionGroup).not.toHaveBeenCalled();
+  });
+});
+
+describe('recomputeAdminAnalytics (scheduled)', () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+    mockFirestoreState.admins = new Set<string>();
+    mockFirestoreState.users = [];
+    mockFirestoreState.dashboards = [];
+    mockFirestoreState.aiUsage = [];
+    mockFirestoreState.docs = new Map();
+    mockFirestoreState.organizations = [];
+  });
+
+  it('writes a snapshot doc only for orgs with status active or trial; skips archived', async () => {
+    // Mix of statuses across three orgs. The scheduler must recompute for
+    // the active + trial ones and explicitly skip the archived org so a
+    // legacy/deactivated tenant doesn't waste compute on every run.
+    mockFirestoreState.organizations = [
+      { id: 'orono', status: 'active' },
+      { id: 'demo-school', status: 'trial' },
+      { id: 'legacy-school', status: 'archived' },
+    ];
+
+    // Seed a single member so `computeAnalyticsForOrg` has a roster to iterate.
+    // Reused across orgs since `mockFirestoreState.users` is global to the
+    // mock; the per-org `members` collection lookup returns the same list,
+    // which is fine for the org-filtering assertion this test is about.
+    mockFirestoreState.users = [
+      {
+        id: 'uid-1',
+        data: {
+          email: 'teacher@orono.k12.mn.us',
+          lastLogin: Date.now(),
+          buildings: [],
+        },
+      },
+    ];
+
+    // The test mock for `firebase-functions/v2/scheduler` returns the inner
+    // handler directly, so the type-system view (a `ScheduleFunction` taking
+    // event + context) doesn't match the actual runtime shape (a thunk).
+    // Cast to a no-arg async function for the call.
+    await (recomputeAdminAnalytics as unknown as () => Promise<void>)();
+
+    // The two active/trial orgs each got a snapshot written; the archived
+    // org did not. Verifies the `ACTIVE_ORG_STATUSES` filter at the top of
+    // the scheduler — a regression that recomputes archived orgs would
+    // immediately get caught here.
+    expect(
+      mockFirestoreState.docs.get('organizations/orono/analytics/snapshot')
+    ).toBeDefined();
+    expect(
+      mockFirestoreState.docs.get(
+        'organizations/demo-school/analytics/snapshot'
+      )
+    ).toBeDefined();
+    expect(
+      mockFirestoreState.docs.get(
+        'organizations/legacy-school/analytics/snapshot'
+      )
+    ).toBeUndefined();
+
+    // Snapshot shape: schemaVersion + computedAt + nextRecomputeAt (24h
+    // delta) + payload. Lock in the contract the reader at
+    // `readAnalyticsSnapshot` enforces.
+    const written = mockFirestoreState.docs.get(
+      'organizations/orono/analytics/snapshot'
+    ) as unknown as AnalyticsSnapshotDoc;
+    expect(written.schemaVersion).toBe(SNAPSHOT_SCHEMA_VERSION);
+    expect(typeof written.computedAt).toBe('number');
+    expect(written.nextRecomputeAt).toBe(
+      written.computedAt + 24 * 60 * 60 * 1000
+    );
+    expect(written.payload.users.total).toBe(1);
+  });
+
+  it('survives a per-org compute failure and continues to the next org', async () => {
+    // Pin the failure: the first org's member-roster read rejects, while
+    // the second org's read succeeds. The scheduler must log the failure,
+    // skip writing a snapshot for the failing org, and still complete the
+    // recompute for the healthy one. Without this resilience, a single
+    // misconfigured org would starve every other tenant of fresh analytics.
+    mockFirestoreState.organizations = [
+      { id: 'broken-org', status: 'active' },
+      { id: 'healthy-org', status: 'active' },
+    ];
+    mockFirestoreState.users = [
+      {
+        id: 'uid-2',
+        data: {
+          email: 'teacher@healthy.org',
+          lastLogin: Date.now(),
+          buildings: [],
+        },
+      },
+    ];
+
+    // Force the first call to `collection('organizations/broken-org/members')`
+    // to reject. Subsequent calls (healthy-org) use the normal mock path.
+    //
+    // Direct property replacement rather than `vi.spyOn` because spying on a
+    // property that's already a `vi.fn()` collapses identities — the captured
+    // "original" reference ends up being the spy itself, and a delegating
+    // impl that calls `original(name)` infinite-loops. Swapping the property
+    // with a fresh wrapper and calling the saved original breaks the cycle.
+    const originalCollection = mockFirestore.collection;
+    const wrappedCollection = vi.fn((name: string) => {
+      if (name === 'organizations/broken-org/members') {
+        return {
+          get: vi.fn(() =>
+            Promise.reject(new Error('synthetic failure: roster read'))
+          ),
+        } as unknown as ReturnType<typeof originalCollection>;
+      }
+      return originalCollection(name);
+    });
+    mockFirestore.collection =
+      wrappedCollection as unknown as typeof originalCollection;
+
+    try {
+      // The test mock for `firebase-functions/v2/scheduler` returns the inner
+      // handler directly, so the type-system view (a `ScheduleFunction` taking
+      // event + context) doesn't match the actual runtime shape (a thunk).
+      // Cast to a no-arg async function for the call.
+      await (recomputeAdminAnalytics as unknown as () => Promise<void>)();
+
+      expect(
+        mockFirestoreState.docs.get(
+          'organizations/broken-org/analytics/snapshot'
+        )
+      ).toBeUndefined();
+      expect(
+        mockFirestoreState.docs.get(
+          'organizations/healthy-org/analytics/snapshot'
+        )
+      ).toBeDefined();
+    } finally {
+      mockFirestore.collection = originalCollection;
+    }
   });
 });
